+++ server/app.py
@@ -26,6 +26,7 @@ from views.bookmarkAPI import Bookmarks3
 
 
 
+from views.modelAPI import AIModel
 import config
 
 app = Flask(__name__)
@@ -42,6 +43,9 @@ api = Api(app)
 
 # add ROOT to PATH
 FILE = Path(__file__).resolve()
+YOLO_ROOT = '/2022-2-SCS4031-9to6/Model/object_detection'
+if YOLO_ROOT not in sys.path:
+    sys.path.append(YOLO_ROOT)
 ROOT = FILE.parents[1]
 if str(ROOT) not in sys.path:
     sys.path.append(str(ROOT))
@@ -71,6 +75,7 @@ api.add_resource(Postings, '/Postings/<posting_index>')
 api.add_resource(PostingList, '/Postings')
 api.add_resource(Login, '/Login')
 api.add_resource(Logout, '/Logout')
+api.add_resource(AIModel, '/inference/<cctv_id>')
 api.add_resource(Bookmarks, '/Bookmark')
 api.add_resource(Bookmarks2, '/Bookmark/<M_ID>/<C_ID>')
 api.add_resource(Bookmarks3, '/Bookmark/<m_id>')
