+++ server/views/historyAPI.py
@@ -10,6 +10,7 @@ import requests
 import json
 from datetime import datetime, date, timedelta
 import math
+from views.utils import s3_upload_file, s3_delete_image
 
 class FloodHistoryList(Resource):
 
@@ -102,33 +103,6 @@ def addFloodHistory(cctvID, stage, imageURL):
     else:
         return False
 
-
-
-# s3에 객체 업로드
-def s3_put_object(s3, bucket, filepath, access_key):
-    """
-    s3 bucket에 지정 파일 업로드
-    :param s3: 연결된 s3 객체(boto3 client)
-    :param bucket: 버킷명
-    :param filepath: 파일 위치
-    :param access_key: 저장 파일명
-    :return: 성공 시 True, 실패 시 False 반환
-    """
-    try:
-        s3.upload_file(
-            Filename=filepath,
-            Bucket=bucket,
-            Key=access_key,
-            ExtraArgs={"ContentType": "image/jpg", "ACL": "public-read"},
-        )
-    except Exception as e:
-        print(e)
-        return False
-    return True
-
-# ret = s3_put_object(s3, '9to6bucket', 'flooding.jpg', 'flooding.jpg')
-# if ret: print('success')
-# else: print('fail')
 
 # 기상청 초단기실황 api
